+++ bpxe/src/data_object.rs
@@ -1,0 +1,3 @@
+//! Data Object
+
+pub trait DataObject<T> {}