+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Inclusive gateway support
+- `scriptTask` activity support
 
 ### Fixed
 
 - Rhai-based condition expressions shouldn't be full scripts
+- Parsing of `script` element (`bpxe-bpmn-schema` crate)
 
 ## [0.1.2] - 2021-01-23